Refactor UI module code for better maintainability
- Refactored index.js by extracting components into separate modules
- Created swipers.js for all carousel functionality
- Created smoothCounters.js for counter animation
- Improved code structure with function early returns
